:truck: s11: Moves Banner component to its own directory
