+++ coffee-app/src/components/Banner/Banner.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import styles from "./banner.module.css";
+import styles from "./Banner.module.css";
 
 import { useEffect } from "react";
 import { useRouter } from "next/navigation";
